Changed Menu.cpp and .h to be Chess.cpp and .h.   Updated Makefile, and fixed Fancy refresh
